pthread: remove enter_critical_section in pthread_mutex
reason:
We would like to replace the critical section with a small lock.
