+++ sched/init/nx_start.c
@@ -469,6 +469,10 @@ static void idle_group_initialize(void)
 
       nxtask_joininit(tcb);
 
+#ifndef CONFIG_PTHREAD_MUTEX_UNSAFE
+      spin_lock_init(&tcb->mutex_lock);
+#endif
+
 #ifdef CONFIG_SMP
       /* Create a stack for all CPU IDLE threads (except CPU0 which already
        * has a stack).
