+++ sched/pthread/CMakeLists.txt
@@ -46,8 +46,7 @@ if(NOT CONFIG_DISABLE_PTHREAD)
       pthread_setschedprio.c)
 
   if(NOT CONFIG_PTHREAD_MUTEX_UNSAFE)
-    list(APPEND SRCS pthread_mutex.c pthread_mutexconsistent.c
-         pthread_mutexinconsistent.c)
+    list(APPEND SRCS pthread_mutex.c pthread_mutexconsistent.c)
   endif()
 
   if(CONFIG_SMP)
