+++ sched/task/task_fork.c
@@ -164,6 +164,10 @@ FAR struct task_tcb_s *nxtask_setup_fork(start_t retaddr)
 
   nxtask_joininit(&child->cmn);
 
+#ifndef CONFIG_PTHREAD_MUTEX_UNSAFE
+  spin_lock_init(&child->cmn.mutex_lock);
+#endif
+
   /* Allocate a new task group with the same privileges as the parent */
 
   ret = group_initialize(child, ttype);
